Using functional enrichment results from gprofiler2 to create a Cytoscape network

User selected enrichment terms are used to create a Cytoscape network where the selected terms and the genes that where part of the enrichment analysis are all represented as nodes. Edges are linking the genes to their terms. The selection of the term can by specifying the source of the terms (GO:MF, REAC, TF, etc...) or by listing the selected term IDs. The network is only generated when there is at least on significant term to graph.

createNetwork(
  gostObject,
  source = c("TERM_ID", "GO:MF", "GO:CC", "GO:BP", "KEGG", "REAC", "TF", "MIRNA", "HPA",
    "CORUM", "HP", "WP"),
  termIDs = NULL,
  removeRoot = TRUE,
  title = "gprofiler network",
  collection = "enrichment results",
  fileName = "gprofilerNetwork.cx"
)

Arguments

gostObject

a list created by gprofiler2 that contains the results from an enrichment analysis.

source

a character string representing the selected source that will be used to generate the network. To hand-pick the terms to be used, "TERM_ID" should be used and the list of selected term IDs should be passed through the termIDs parameter. The possible sources are "GO:BP" for Gene Ontology Biological Process, "GO:CC" for Gene Ontology Cellular Component, "GO:MF" for Gene Ontology Molecular Function, "KEGG" for Kegg, "REAC" for Reactome, "TF" for TRANSFAC, "MIRNA" for miRTarBase, "CORUM" for CORUM database, "HP" for Human phenotype ontology and "WP" for WikiPathways. Default: "TERM_ID".

termIDs

a vector of character strings that contains the term IDS retained for the creation of the network. Default: NULL.

removeRoot

a logical that specified if the root terms of the selected source should be removed (when present). Default: TRUE.

title

a character string representing the name assigned to the network. Default: "gprofiler network".

collection

a character string representing the collection name assigned to the network. Default: "enrichment results".

fileName

a character string representing the name of the CX JSON file that is created when Cytoscape is not running. The name must have a '.cx' extension. Default: "gprofilerNetwork_01.cx".

Value

TRUE

Author

Astrid Deschênes

Examples

## Loading dataset containing result from an enrichment analysis done with
## gprofiler2
data(demoGOST)

if (FALSE) {

## Create network for Gene Ontology - Molecular Function related results
createNetwork(gostObject=demoGOST, source="GO:MF", removeRoot=FALSE,
    title="GO Molecular Function Graph")

}